BeatDownloader 1.0.2
Changed method for getting the Beat Saber install directory
Added more messages for a better flow
Changed xpath for map keys

+++ BeatDownloader.bat
@@ -1,25 +1,32 @@
 @echo off
 SetLocal EnableDelayedExpansion
 
-if not exist "BeatPath.txt" (
-   type nul > BeatPath.txt
+where xidel 2>nul || echo "Xidel needs to be added to the current directory in order to continue" && goto :End
+if exist "response.json" (
+   echo A partial run has been detected && echo 1^) Continue the previous run && echo 2^) Start over
+   choice /c 12 /m "Please select an option"
+   if errorlevel 2 goto :Start
+   if errorlevel 1 goto :Continue
 )
-set beatSaberPath = ""
-for /f "tokens=*" %%a in (BeatPath.txt) do set beatSaberPath=%%a
-set "psCommand="(new-object -COM 'Shell.Application')^
-.BrowseForFolder(0,'Please choose your BeatSaber folder.',0,0).self.path""
-if "%beatSaberPath%"=="" (
-   for /f "usebackq delims=" %%b in (`powershell %psCommand%`) do set "beatSaberPath=%%b"
-   echo|set /p ="!beatSaberPath!">BeatPath.txt
-)
-if [!beatSaberPath!]==[] (
-   echo "A directory must be selected to continue"
-   goto :End
-)
+
+:Start
+curl -o response.json https://beat-savior.herokuapp.com/api/maps/ranked
+
+:Continue
+set startDir=%cd%
+cd "%LocalAppData%\ModAssistant"
+for /f "delims=" %%a in ('dir /S /B user.config') do set modSettings=%%a
+cd %startDir%
+
+xidel -s %modSettings% --xpath './/configuration/userSettings/ModAssistant.Properties.Settings/setting[@name="OCIWindow"]/value' > output.txt
+for /f "tokens=*" %%b in (output.txt) do set ociWindow=%%b
+if "!ociWindow!"=="Yes" echo It is recommended to change the "Show OneClick Installer Window" setting from Yes to Close or No to avoid multiple open installer window processes && echo Press Y to continue anyway . . . && choice /c y /n
+
+xidel -s %modSettings% --xpath './/configuration/userSettings/ModAssistant.Properties.Settings/setting[@name="InstallFolder"]' > output.txt
+for /f "tokens=*" %%b in (output.txt) do set beatSaberPath=%%b
 echo Using path: !beatSaberPath!
 
-curl -o response.json https://beat-savior.herokuapp.com/api/maps/ranked
-xidel response.json --xpath ".//maps/Key" > output.txt
+xidel response.json --xpath ".//key" > output.txt
 for /f %%c in ('Find "" /v /c ^< output.txt') do set total=%%c
 set curr=0
 for /f "tokens=*" %%d in (output.txt) do (
@@ -28,7 +35,7 @@ for /f "tokens=*" %%d in (output.txt) do (
    ) else (
       echo "Saving new custom level: %%d"
       start beatsaver://%%d
-      timeout 10 >nul
+      timeout 8 >nul
    )
    set /a "curr+=1"
    echo Completed: !curr! / %total%
